+++ lib/commanded/process_managers/process_router.ex
@@ -22,6 +22,7 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
       :command_dispatcher,
       :consistency,
       :event_timeout,
+      :idle_timeout,
       :process_manager_name,
       :process_manager_module,
       :subscribe_from,
@@ -45,7 +46,8 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
       command_dispatcher: command_dispatcher,
       consistency: opts[:consistency] || :eventual,
       subscribe_from: opts[:start_from] || :origin,
-      event_timeout: opts[:event_timeout]
+      event_timeout: opts[:event_timeout],
+      idle_timeout: opts[:idle_timeout] || :infinity
     }
 
     Registration.start_link(name, __MODULE__, state)
@@ -156,9 +158,7 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
   def handle_info({:subscribed, subscription}, %State{subscription: subscription} = state) do
     Logger.debug(fn -> describe(state) <> " has successfully subscribed to event store" end)
 
-    %State{command_dispatcher: command_dispatcher} = state
-
-    {:ok, supervisor} = Supervisor.start_link(command_dispatcher, self())
+    {:ok, supervisor} = Supervisor.start_link()
 
     {:noreply, %State{state | supervisor: supervisor}}
   end
@@ -435,19 +435,24 @@ defmodule Commanded.ProcessManagers.ProcessRouter do
 
   defp start_process_manager(process_uuid, %State{} = state) do
     %State{
+      command_dispatcher: command_dispatcher,
+      idle_timeout: idle_timeout,
       process_managers: process_managers,
       process_manager_name: process_manager_name,
       process_manager_module: process_manager_module,
       supervisor: supervisor
     } = state
 
-    {:ok, process_manager} =
-      Supervisor.start_process_manager(
-        supervisor,
-        process_manager_name,
-        process_manager_module,
-        process_uuid
-      )
+    opts = [
+      command_dispatcher: command_dispatcher,
+      idle_timeout: idle_timeout,
+      process_manager_name: process_manager_name,
+      process_manager_module: process_manager_module,
+      process_router: self(),
+      process_uuid: process_uuid
+    ]
+
+    {:ok, process_manager} = Supervisor.start_process_manager(supervisor, opts)
 
     Process.monitor(process_manager)